🔨 Refactor: member package entity들 refactoring
