+++ src/main/java/com/cheeup/domain/member/MemberSkill.java
@@ -1,0 +1,16 @@
+package com.cheeup.domain.member;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Table(name = "member_skills")
+public class MemberSkill {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+}
